requirements.txt
streamlit==1.36.0
Pillow==10.4.0